requirements.txt
blinker==1.7.0
click==8.1.7
Cython==3.0.6
Flask==3.0.0
Flask-Cors==4.0.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.2.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
patsy==0.5.4
pmdarima==2.0.4
protobuf==4.21.12
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
SQLAlchemy==2.0.23
statsmodels==0.14.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
